+++ Assets/Scripts/Player Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 17d636ce6adf17940bb082d23d30f0c1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
